refactor: move annotate_user_interactions to stories services; clean test imports
Move annotate_user_interactions from apps/stories/views.py to
apps/stories/services.py so cross-app callers (users/views.py)
import from the service layer rather than another app's views.

Also move datetime, timezone, and SavedStory imports from inside
test methods to file-level imports in users/tests/test_services.py
and users/tests/test_views.py.
